XKeys can now be used with data types other than array or hash (as long as they support array- or hash-like interfaces).
New custom nodes can be generated by defining an #xkeys_new
method. The method is responsible for adding XKeys extensions to
the new node as needed/desired.

See Sarah::XK in gem sarah-xk for an example.

Using :[] (push mode) as the first of multiple keys now works
correctly. (#push is used to add :[]-keyed values so as not to
depend on #size for the correct index.)